+++ zebrad/src/commands/revhex.rs
@@ -5,6 +5,20 @@
 use abscissa_core::{Command, Options, Runnable};
 use std::io::stdin;
 
+/// Returns the hexadecimal-encoded string `s` in byte-reversed order.
+pub fn byte_reverse_hex(s: &str) -> String {
+    String::from_utf8(
+        s.as_bytes()
+            .chunks(2)
+            .rev()
+            .map(|c| c.iter())
+            .flatten()
+            .cloned()
+            .collect::<Vec<u8>>(),
+    )
+    .expect("input should be ascii")
+}
+
 /// `revhex` subcommand
 #[derive(Command, Debug, Default, Options)]
 pub struct RevhexCmd {
@@ -26,10 +40,10 @@ impl Runnable for RevhexCmd {
             // We can distinguish EOF from an empty line, because the newline is
             // included in the buffer, so empty lines return Ok(1).
             while stdin().read_line(&mut input).unwrap_or(0) > 0 {
-                println!("{}", zebra_chain::utils::byte_reverse_hex(&input.trim()));
+                println!("{}", byte_reverse_hex(&input.trim()));
             }
         } else {
-            println!("{}", zebra_chain::utils::byte_reverse_hex(&self.input));
+            println!("{}", byte_reverse_hex(&self.input));
         }
     }
 }
